requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosmtplib==3.0.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
backports.asyncio.runner==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
cachetools==6.1.0
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorlog==6.9.0
contourpy==1.3.2
coverage==7.10.1
cryptography==45.0.5
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
easyocr==1.7.2
ecdsa==0.19.1
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.58.2
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
GPUtil==1.4.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.1
idna==3.7
imageio==2.37.0
imagesize==1.4.1
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kombu==5.5.4
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.67
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
ollama==0.5.1
openai==1.93.0
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
packaging==24.2
paddleocr==3.1.0
paddlepaddle==3.1.0
pandas==2.3.0
passlib==1.7.4
pillow==11.2.1
pillow-avif-plugin==1.5.2
pillow_heif==0.22.0
pluggy==1.6.0
premailer==3.10.0
prettytable==3.16.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.1
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
roboflow==1.2.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
sendgrid==6.12.4
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.47.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.153
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
zstandard==0.23.0
